Added Track Generator
Added a script that generates a track. For the moment, the script always generates the same track based on a seed and user parameters, as it is important that competing players are subject to the same variables to make sure the competition is safe. Also wired in all the buttons so users can navigate between the menu and different scenes properly