StartPanel + Score Fix
Added a panel at the start that askes the user to press W to start. Only then will the timer start and the hovercraft start moving.

Amended the FinishLine behaviour to stop the hovercraft moving when the user hits the finish line. Also fixed the final score which was previously using the time of when the score submit button was pressed rather than when the finish line was crossed